Resolve Fix

+++ src/main/java/com/example/springexampractice/controller/EmployeeController.java
@@ -34,7 +34,7 @@ public class EmployeeController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Employee> update(@RequestBody Employee employee, @PathVariable Long id) {
+    public ResponseEntity<Employee> update(@RequestBody Employee employee, @PathVariable Integer id) {
         return ResponseEntity.ok(employeeService.update(employee, id));
     }
